+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Next Release]
+- null check for issue #475
+
 ## 2.5.0 - 2021-2-24
 - Published to MavenCentral
 - Removed dependencies from jCenter