+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Next Release]
+
+## 2.5.1 - 2021-4-10
 - null check for issue #475
 - respect custom master list when setting auto detected country #454
 - update gradle plug-in